074/requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
olefile==0.44
Pillow==4.1.1
requests==2.17.3
urllib3==1.21.1
